Modified MemberScreen to support JSC and USFSA memberships.
Both membership types are now listed on the Membership page, and each
item has its own label and cost.  So the display is working, but the
backend for calculating the totals, registering the memberships,
building the paypal form, and charging the correct amounts has not been
implemented yet.  This is a checkpoint checkin.